# fix: cache source maps to speed up setting