The old debug adapter took about ~100ms to set a breakpoint. The new one
was taking ~300ms. It turns out the first time you try to resolve a
breakpoint location with a parsed sourcemap, it takes roughly ~200ms,
and roughly ~0ms after that point. We weren't caching sourcemaps,
reparsing them every time we set a breakpoint. The math adds up :)

Fixes #178